refactor doInitialPcbTraceRender to separate file
