+++ lib/components/primitive-components/Trace/Trace.ts
@@ -46,26 +46,6 @@ import { getObstaclesFromCircuitJson } from "lib/utils/obstacles/getObstaclesFro
 import { computeSchematicNetLabelCenter } from "lib/utils/schematic/computeSchematicNetLabelCenter"
 import { Trace_doInitialSchematicTraceRender } from "./Trace_doInitialSchematicTraceRender"
 
-type PcbRouteObjective =
-  | RouteHintPoint
-  | {
-      layers: string[]
-      x: number
-      y: number
-      via?: boolean
-      pcb_port_id?: string
-    }
-
-const portToObjective = (port: Port): PcbRouteObjective => {
-  const portPosition = port._getGlobalPcbPositionAfterLayout()
-  return {
-    ...portPosition,
-    layers: port.getAvailablePcbLayers(),
-  }
-}
-
-const SHOULD_USE_SINGLE_LAYER_ROUTING = false
-
 export class Trace
   extends PrimitiveComponent<typeof traceProps>
   implements TraceI
@@ -374,352 +354,9 @@ export class Trace
       })
     }
   }
+
   doInitialPcbTraceRender(): void {
-    if (this.root?.pcbDisabled) return
-    const { db } = this.root!
-    const { _parsedProps: props, parent } = this
-    const subcircuit = this.getSubcircuit()
-
-    if (!parent) throw new Error("Trace has no parent")
-
-    if (subcircuit._parsedProps.routingDisabled) {
-      return
-    }
-
-    // Check for cached route
-    const cachedRoute = subcircuit._parsedProps.pcbRouteCache?.pcbTraces
-    if (cachedRoute) {
-      const pcb_trace = db.pcb_trace.insert({
-        route: cachedRoute.flatMap((trace) => trace.route),
-        source_trace_id: this.source_trace_id!,
-        subcircuit_id: subcircuit?.subcircuit_id ?? undefined,
-        pcb_group_id: this.getGroup()?.pcb_group_id ?? undefined,
-      })
-      this.pcb_trace_id = pcb_trace.pcb_trace_id
-      return
-    }
-
-    if (!subcircuit._shouldUseTraceByTraceRouting()) {
-      return
-    }
-
-    const { allPortsFound, ports } = this._findConnectedPorts()
-    const portsConnectedOnPcbViaNet: Port[] = []
-
-    if (!allPortsFound) return
-
-    const portsWithoutMatchedPcbPrimitive: Port[] = []
-    for (const port of ports) {
-      if (!port._hasMatchedPcbPrimitive()) {
-        portsWithoutMatchedPcbPrimitive.push(port)
-      }
-    }
-
-    if (portsWithoutMatchedPcbPrimitive.length > 0) {
-      db.pcb_trace_error.insert({
-        error_type: "pcb_trace_error",
-        source_trace_id: this.source_trace_id!,
-        message: `Some ports did not have a matching PCB primitive (e.g. a pad or plated hole), this can happen if a footprint is missing. As a result, ${this} wasn't routed. Missing ports: ${portsWithoutMatchedPcbPrimitive.map((p) => p.getString()).join(", ")}`,
-        pcb_trace_id: this.pcb_trace_id!,
-        pcb_component_ids: [],
-        pcb_port_ids: portsWithoutMatchedPcbPrimitive
-          .map((p) => p.pcb_port_id!)
-          .filter(Boolean),
-      })
-      return
-    }
-
-    const nets = this._findConnectedNets().netsWithSelectors
-
-    if (ports.length === 0 && nets.length === 2) {
-      // Find the two optimal points to connect the two nets
-      this.renderError(
-        `Trace connects two nets, we haven't implemented a way to route this yet`,
-      )
-      return
-      // biome-ignore lint/style/noUselessElse: <explanation>
-    } else if (ports.length === 1 && nets.length === 1) {
-      // Add a port from the net that is closest to the port
-      const port = ports[0]
-      const portsInNet = nets[0].net.getAllConnectedPorts()
-      const otherPortsInNet = portsInNet.filter((p) => p !== port)
-      if (otherPortsInNet.length === 0) {
-        console.log(
-          "Nothing to connect this port to, the net is empty. TODO should emit a warning!",
-        )
-        return
-      }
-      const closestPortInNet = getClosest(port, otherPortsInNet)
-
-      portsConnectedOnPcbViaNet.push(closestPortInNet)
-
-      ports.push(closestPortInNet)
-    } else if (ports.length > 1 && nets.length >= 1) {
-      this.renderError(
-        `Trace has more than one port and one or more nets, we don't currently support this type of complex trace routing`,
-      )
-      return
-    }
-
-    const hints = ports.flatMap((port) =>
-      port.matchedComponents.filter((c) => c.componentName === "TraceHint"),
-    ) as TraceHint[]
-
-    const pcbRouteHints = (this._parsedProps.pcbRouteHints ?? []).concat(
-      hints.flatMap((h) => h.getPcbRouteHints()),
-    )
-
-    if (ports.length > 2) {
-      this.renderError(
-        `Trace has more than two ports (${ports
-          .map((p) => p.getString())
-          .join(
-            ", ",
-          )}), routing between more than two ports for a single trace is not implemented`,
-      )
-      return
-    }
-
-    const alreadyRoutedTraces = this.getSubcircuit()
-      .selectAll("trace")
-      .filter(
-        (trace) => trace.renderPhaseStates.PcbTraceRender.initialized,
-      ) as Trace[]
-
-    const isAlreadyRouted = alreadyRoutedTraces.some(
-      (trace) =>
-        trace._portsRoutedOnPcb.length === ports.length &&
-        trace._portsRoutedOnPcb.every((portRoutedByOtherTrace) =>
-          ports.includes(portRoutedByOtherTrace),
-        ),
-    )
-
-    if (isAlreadyRouted) {
-      return
-    }
-
-    let orderedRouteObjectives: PcbRouteObjective[] = []
-
-    if (pcbRouteHints.length === 0) {
-      orderedRouteObjectives = [
-        portToObjective(ports[0]),
-        portToObjective(ports[1]),
-      ]
-    } else {
-      // When we have hints, we have to order the hints then route between each
-      // terminal of the trace and the hints
-      // TODO order based on proximity to ports
-      orderedRouteObjectives = [
-        portToObjective(ports[0]),
-        ...pcbRouteHints,
-        portToObjective(ports[1]),
-      ]
-    }
-
-    // Hints can indicate where there should be a via, but the layer is allowed
-    // to be unspecified, therefore we need to find possible layer combinations
-    // to go to each hint and still route to the start and end points
-    const candidateLayerCombinations = findPossibleTraceLayerCombinations(
-      orderedRouteObjectives,
-    )
-
-    if (
-      SHOULD_USE_SINGLE_LAYER_ROUTING &&
-      candidateLayerCombinations.length === 0
-    ) {
-      this.renderError(
-        `Could not find a common layer (using hints) for trace ${this.getString()}`,
-      )
-      return
-    }
-
-    const connMap = getFullConnectivityMapFromCircuitJson(
-      this.root!.db.toArray(),
-    )
-
-    // Cache the PCB obstacles, they'll be needed for each segment between
-    // ports/hints
-    const [obstacles, errGettingObstacles] = tryNow(
-      () => getObstaclesFromCircuitJson(this.root!.db.toArray() as any), // Remove as any when autorouting-dataset gets updated
-    )
-
-    if (errGettingObstacles) {
-      this.renderError({
-        type: "pcb_trace_error",
-        error_type: "pcb_trace_error",
-        pcb_trace_error_id: this.pcb_trace_id!,
-        message: `Error getting obstacles for autorouting: ${errGettingObstacles.message}`,
-        source_trace_id: this.source_trace_id!,
-        center: { x: 0, y: 0 },
-        pcb_port_ids: ports.map((p) => p.pcb_port_id!),
-        pcb_trace_id: this.pcb_trace_id!,
-        pcb_component_ids: [],
-      })
-      return
-    }
-
-    for (const obstacle of obstacles) {
-      const connectedTo = obstacle.connectedTo
-      if (connectedTo.length > 0) {
-        const netId = connMap.getNetConnectedToId(obstacle.connectedTo[0])
-        if (netId) {
-          obstacle.connectedTo.push(netId)
-        }
-      }
-    }
-
-    let orderedRoutePoints: PcbRouteObjective[] = []
-    if (candidateLayerCombinations.length === 0) {
-      orderedRoutePoints = orderedRouteObjectives
-    } else {
-      // TODO explore all candidate layer combinations if one fails
-      const candidateLayerSelections = candidateLayerCombinations[0].layer_path
-
-      /**
-       * Apply the candidate layer selections to the route objectives, now we
-       * have a set of points that have definite layers
-       */
-      orderedRoutePoints = orderedRouteObjectives.map((t, idx) => {
-        if (t.via) {
-          return {
-            ...t,
-            via_to_layer: candidateLayerSelections[idx],
-          }
-        }
-        return { ...t, layers: [candidateLayerSelections[idx]] }
-      })
-    }
-    ;(orderedRoutePoints[0] as any).pcb_port_id = ports[0].pcb_port_id
-    ;(orderedRoutePoints[orderedRoutePoints.length - 1] as any).pcb_port_id =
-      ports[1].pcb_port_id
-
-    const routes: PcbTrace["route"][] = []
-    for (const [a, b] of pairs(orderedRoutePoints)) {
-      const dominantLayer =
-        "via_to_layer" in a ? (a.via_to_layer as LayerRef) : null
-      const BOUNDS_MARGIN = 2 //mm
-
-      const aLayer =
-        "layers" in a && a.layers.length === 1
-          ? a.layers[0]
-          : (dominantLayer ?? "top")
-      const bLayer =
-        "layers" in b && b.layers.length === 1
-          ? b.layers[0]
-          : (dominantLayer ?? "top")
-
-      const pcbPortA = "pcb_port_id" in a ? a.pcb_port_id : null
-      const pcbPortB = "pcb_port_id" in b ? b.pcb_port_id : null
-
-      const minTraceWidth =
-        this.getSubcircuit()._parsedProps.minTraceWidth ?? 0.16
-
-      const ijump = new MultilayerIjump({
-        OBSTACLE_MARGIN: minTraceWidth * 2,
-        isRemovePathLoopsEnabled: true,
-        optimizeWithGoalBoxes: Boolean(pcbPortA && pcbPortB),
-        connMap,
-        input: {
-          obstacles,
-          minTraceWidth,
-          connections: [
-            {
-              name: this.source_trace_id!,
-              pointsToConnect: [
-                { ...a, layer: aLayer, pcb_port_id: pcbPortA! },
-                { ...b, layer: bLayer, pcb_port_id: pcbPortB! },
-              ],
-            },
-          ],
-          layerCount: 2,
-          bounds: {
-            minX: Math.min(a.x, b.x) - BOUNDS_MARGIN,
-            maxX: Math.max(a.x, b.x) + BOUNDS_MARGIN,
-            minY: Math.min(a.y, b.y) - BOUNDS_MARGIN,
-            maxY: Math.max(a.y, b.y) + BOUNDS_MARGIN,
-          },
-        },
-      })
-      let traces: SimplifiedPcbTrace[] | null = null
-      try {
-        traces = ijump.solveAndMapToTraces()
-      } catch (e: any) {
-        this.renderError({
-          type: "pcb_trace_error",
-          pcb_trace_error_id: this.source_trace_id!,
-          error_type: "pcb_trace_error",
-          message: `error solving route: ${e.message}`,
-          source_trace_id: this.pcb_trace_id!,
-          center: { x: (a.x + b.x) / 2, y: (a.y + b.y) / 2 },
-          pcb_port_ids: ports.map((p) => p.pcb_port_id!),
-          pcb_trace_id: this.pcb_trace_id!,
-          pcb_component_ids: ports.map((p) => p.pcb_component_id!),
-        })
-      }
-      if (!traces) return
-      if (traces.length === 0) {
-        this.renderError({
-          type: "pcb_trace_error",
-          error_type: "pcb_trace_error",
-          pcb_trace_error_id: this.pcb_trace_id!,
-          message: `Could not find a route for ${this}`,
-          source_trace_id: this.source_trace_id!,
-          center: { x: (a.x + b.x) / 2, y: (a.y + b.y) / 2 },
-          pcb_port_ids: ports.map((p) => p.pcb_port_id!),
-          pcb_trace_id: this.pcb_trace_id!,
-          pcb_component_ids: ports.map((p) => p.pcb_component_id!),
-        })
-        return
-      }
-      const [trace] = traces as PcbTrace[]
-
-      // If the autorouter didn't specify a layer, use the dominant layer
-      // Some of the single-layer autorouters don't add the layer property
-      if (dominantLayer) {
-        trace.route = trace.route.map((p) => {
-          if (p.route_type === "wire" && !p.layer) {
-            p.layer = dominantLayer
-          }
-          return p
-        })
-      }
-
-      if (pcbPortA && trace.route[0].route_type === "wire") {
-        trace.route[0].start_pcb_port_id = pcbPortA
-      }
-      const lastRoutePoint = trace.route[trace.route.length - 1]
-      if (pcbPortB && lastRoutePoint.route_type === "wire") {
-        lastRoutePoint.end_pcb_port_id = pcbPortB
-      }
-      routes.push(trace.route)
-    }
-    const mergedRoute = mergeRoutes(routes)
-
-    const traceLength = getTraceLength(mergedRoute)
-    const pcb_trace = db.pcb_trace.insert({
-      route: mergedRoute,
-      source_trace_id: this.source_trace_id!,
-      subcircuit_id: this.getSubcircuit()?.subcircuit_id!,
-      trace_length: traceLength,
-    })
-    this._portsRoutedOnPcb = ports
-    this.pcb_trace_id = pcb_trace.pcb_trace_id
-
-    for (const point of mergedRoute) {
-      if (point.route_type === "via") {
-        db.pcb_via.insert({
-          pcb_trace_id: pcb_trace.pcb_trace_id,
-          x: point.x,
-          y: point.y,
-          hole_diameter: 0.3,
-          outer_diameter: 0.6,
-          layers: [point.from_layer as LayerRef, point.to_layer as LayerRef],
-          from_layer: point.from_layer as LayerRef,
-          to_layer: point.to_layer as LayerRef,
-        })
-      }
-    }
-    this._insertErrorIfTraceIsOutsideBoard(mergedRoute, ports)
+    // TODO
   }
 
   _doInitialSchematicTraceRenderWithDisplayLabel(): void {
